Change filename format & remove old resizing code
Removing the serial number from the start of the filenames because it was a pain to have to export all images in Lightroom in order to update just one file.  Adding "-small.jpg" to make sure the source image has a different name from the processed image.

Removing the resizing code since I always resize in Lightroom now.